LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/animation-display-lock-expected.txt


FAIL Animation events do not fire for a CSS animation running in a display locked subtree assert_false: Animation events do not fire while the animation is running in a display locked subtree expected false got true
FAIL The finished promise does not resolve due to the normal passage of time  for a CSS animation in a display locked subtree assert_equals: expected false but got true
FAIL The finished promise does not resolve due to the normal passage of time  for a CSS transition in a display locked subtree assert_equals: expected false but got true
PASS Events and promises are handled normally for animations without an owning element
